Updated the dequeue function and added dependencies
Updated the dequeue function so that it dequeues elements as quick as possible without a wait. Also fixed the time left to match the table in the front panel. Also added files that are necessary to run the reactor code so that new users do not have to go through hassle of downloading dependencies before running the code.